:recycle: (1002): extract the deliverable card actions and name its footer states
