+++ mcr-frontend/src/components/meeting/DeliverableTypeCard.vue
@@ -17,94 +17,17 @@
         />
       </div>
 
-      <div class="flex items-center gap-1">
-        <template v-if="hasError">
-          <span class="deliverable-tag bg-error-950 text-error-425">
-            <span
-              class="fr-icon-error-fill fr-icon--sm"
-              aria-hidden="true"
-            />
-            {{ errorTagLabel }}
-          </span>
-          <DsfrButton
-            v-if="canRegenerate"
-            icon="fr-icon-refresh-line"
-            icon-only
-            no-outline
-            tertiary
-            :title="$t('meeting-v2.deliverable-card.actions.regenerate')"
-            @click="onAction"
-          />
-        </template>
-
-        <div
-          v-else-if="isAvailable"
-          class="flex items-center gap-1"
-        >
-          <DsfrButton
-            v-if="isCustom"
-            icon="fr-icon-refresh-line"
-            icon-only
-            no-outline
-            tertiary
-            :title="$t('meeting-v2.deliverable-card.actions.regenerate')"
-            @click="$emit('customize')"
-          />
-          <a
-            v-if="deliverable?.external_url && isFichierEnabled"
-            :href="deliverable.external_url"
-            target="_blank"
-            rel="noopener noreferrer"
-            class="fr-btn fr-btn--tertiary-no-outline fr-icon-eye-line"
-            :title="$t('meeting-v2.deliverable-card.actions.open-external')"
-          >
-            <span class="sr-only">{{
-              $t('meeting-v2.deliverable-card.actions.open-external')
-            }}</span>
-          </a>
-          <DsfrButton
-            icon="fr-icon-download-line"
-            icon-only
-            no-outline
-            tertiary
-            :title="$t('meeting-v2.deliverable-card.actions.download')"
-            @click="() => deliverable && $emit('download', deliverable.id)"
-          />
-        </div>
-
-        <span
-          v-else-if="isWaiting"
-          class="deliverable-tag bg-[var(--grey-925-125)] text-[var(--text-mention-grey)]"
-        >
-          <span
-            class="fr-icon-time-line fr-icon--sm"
-            aria-hidden="true"
-          />
-          {{ $t('meeting-v2.deliverable-card.tag.waiting') }}
-        </span>
-
-        <span
-          v-else-if="isLoading"
-          class="deliverable-tag bg-info-950 text-info-425"
-        >
-          <VIcon
-            name="ri-loader-3-line"
-            animation="spin"
-            :scale="0.9"
-          />
-          {{ $t('meeting-v2.deliverable-card.tag.in-progress') }}
-        </span>
-
-        <DsfrButton
-          v-else-if="canGenerate"
-          icon="fr-icon-sparkling-2-line"
-          size="sm"
-          class="generate-button"
-          @click="onAction"
-        >
-          {{ actionLabel }}
-        </DsfrButton>
-      </div>
+      <DeliverableTypeCardActions
+        :state="state"
+        :deliverable="deliverable"
+        :is-custom="isCustom"
+        :can-regenerate="canRegenerate"
+        :error-tag-label="errorTagLabel"
+        :action-label="actionLabel"
+        @action="onAction"
+        @customize="$emit('customize')"
+        @download="(id) => $emit('download', id)"
+      />
     </div>
   </div>
 </template>
@@ -114,6 +37,8 @@ import { useFeatureFlag } from '@/composables/use-feature-flag';
 import { t } from '@/plugins/i18n';
 import type { DeliverableDto, DeliverableType } from '@/services/deliverables/deliverables.types';
 import DeliverableFeedbackThumbs from './deliverable-feedback/DeliverableFeedbackThumbs.vue';
+import DeliverableTypeCardActions from './DeliverableTypeCardActions.vue';
+import type { DeliverableCardState } from './deliverable-card-state';
 
 const props = defineProps<{
   type: DeliverableType;
@@ -129,7 +54,6 @@ const emit = defineEmits<{
   download: [id: number];
 }>();
 
-const isFichierEnabled = useFeatureFlag('fichier-integration');
 const isDeliverableFeedbackEnabled = useFeatureFlag('deliverable-feedback');
 
 const TYPE_KEY_MAP: Record<DeliverableType, string> = {
@@ -177,6 +101,15 @@ const canRegenerate = computed(
   () => hasError.value && !isTranscription.value && !isTranscriptionFailure.value,
 );
 
+const state = computed<DeliverableCardState>(() => {
+  if (hasError.value) return 'error';
+  if (isAvailable.value) return 'available';
+  if (isWaiting.value) return 'waiting';
+  if (isLoading.value) return 'loading';
+  if (canGenerate.value) return 'generate';
+  return 'none';
+});
+
 const title = computed(() =>
   t(`meeting-v2.deliverable-card.type.${TYPE_KEY_MAP[props.type]}.title`),
 );
@@ -214,27 +147,5 @@ function onAction(): void {
   flex-direction: column;
   gap: 0.5rem;
   background-color: var(--grey-1000-50);
-}
-
-.deliverable-tag {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.25rem;
-  border-radius: 9999px;
-  padding: 0.1875rem 0.625rem;
-  font-size: 0.6875rem;
-  font-weight: 700;
-  text-transform: uppercase;
-  letter-spacing: 0.02em;
-}
-
-.generate-button {
-  border-radius: 9999px !important;
-  box-shadow: 0 2px 6px rgba(0, 0, 40, 0.16);
-}
-
-.deliverable-type-card a.fr-btn::after {
-  display: none;
-  content: none;
 }
 </style>
