+++ mcr-frontend/src/components/meeting/deliverable-card-state.ts
@@ -1,0 +1,7 @@
+export type DeliverableCardState =
+  | 'error'
+  | 'available'
+  | 'waiting'
+  | 'loading'
+  | 'generate'
+  | 'none';
